requirements.txt
py-clob-client>=0.18.0
aiohttp>=3.9.0
pandas>=2.0.0
python-dotenv>=1.0.0
websockets>=12.0
